requirements.txt
streamlit==1.25.0
pyrebase4
firebase-admin==6.1.0
streamlit-cookies-manager
